update build st file

Created-by: y30038433
Commit-by: y30038433
Merged-by: zhangjian_hz11
Description: update build st file

See merge request: cann/pto-tile-lib!54

+++ test/davinci/v310/src/st/testcase/textract/textract_kernel.cpp
@@ -7,7 +7,7 @@ using namespace pto;
 template <typename T>
 __aicore__ inline void DynGM2L1(__cbuf__ T *dst, __gm__ T *src, unsigned TShape0, unsigned TShape1)
 {
-    uint16_t lenBurst = TShape0 * TShape1 * sizeof(T);
+    uint32_t lenBurst = TShape0 * TShape1 * sizeof(T);
     copy_gm_to_cbuf_align_v2(dst, src, 0, 1, lenBurst, 0, 0, 0, 0, 0, 0);
 }
